raft: Move offset_monitor to utils/
The offset_monitor will be used in cloud topics subsystem in a followup
commits.
