+++ src/v/utils/BUILD
@@ -687,3 +687,16 @@ redpanda_cc_library(
         "@seastar",
     ],
 )
+
+redpanda_cc_library(
+    name = "offset_monitor",
+    hdrs = [
+        "offset_monitor.h",
+    ],
+    deps = [
+        "//src/v/base",
+        "//src/v/model",
+        "@abseil-cpp//absl/container:btree",
+        "@seastar",
+    ],
+)
